Update NumPy.py
defined datatype for python array

+++ NumPy.py
@@ -5,7 +5,7 @@
 
 
 import numpy as np
-a=np.array([1,2,3,4,5])
+a=np.array([1,2,3,4,5], dtype=int)
 print(a)
